Generate the Forecast controller #index view and style the UI (#3)
Use the rails generator to create the controller and corresponding view.
Since the application still has no need for a database, there's no need
to generate a full scaffold or model. That might change in the future so
leave the database in for now.

I took UI inspiration from 
[WeatherAPI's own forecasting site](https://www.weatherapi.com/weather/).
If I had more time I would make the background gradient unique for each
weather type. For this commit the data is hardcoded, getting real data
is a future commit. I'm actually pleased with how this app looks. It's
one of my better works. The weather icons come from the API so that
should be easy to pull in a future commit. The one Bootstrap icon is
hardcoded into the DOM, if I find I want another icon then I'll pull the
CDN and properly get the icons loaded in. In the interest of time I left
it in the DOM.